requirements.txt
flask
Pillow